# application/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "OpenReq Requirement Prioritization Recommendation Service"
  version: "1.0.0"
  title: "OpenReq Requirement Prioritization Recommendation Service"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "rsamer@ist.tugraz.at"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:9002"
basePath: "/"
tags:
- name: "recommend"
  description: "OpenReq Requirement Dependency Recommendation Service"
schemes:
- "http"
paths:
  /prioritizer/compute:
    post:
      tags:
      - "Recommender"
      summary: "Retrieve a list of prioritized issues."
      description: ""
      operationId: "recommend_prioritized_issues"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Request object"
        required: true
        schema:
          $ref: "#/definitions/PrioritizedRecommendationsRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/PrioritizedRecommendationsResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "application.controllers.recommendation_controller"
  /prioritizer/chart:
    post:
      tags:
      - "Recommender"
      summary: "Retrieve the chart URL."
      description: ""
      operationId: "generate_chart_url"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Request object"
        required: true
        schema:
          $ref: "#/definitions/ChartRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ChartResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "application.controllers.recommendation_controller"
  /prioritizer/like:
    post:
      tags:
      - "Recommender"
      summary: "Like a requirement."
      description: ""
      operationId: "like_prioritized_issue"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Request object"
        required: true
        schema:
          $ref: "#/definitions/LikeRequirementRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/LikeRequirementResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "application.controllers.recommendation_controller"
  /prioritizer/unlike:
    post:
      tags:
      - "Recommender"
      summary: "Unlike a requirement."
      description: ""
      operationId: "unlike_prioritized_issue"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Request object"
        required: true
        schema:
          $ref: "#/definitions/LikeRequirementRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/LikeRequirementResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "application.controllers.recommendation_controller"
  /prioritizer/dislike:
    post:
      tags:
      - "Recommender"
      summary: "Dislike a requirement."
      description: ""
      operationId: "dislike_prioritized_issue"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Request object"
        required: true
        schema:
          $ref: "#/definitions/LikeRequirementRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/LikeRequirementResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "application.controllers.recommendation_controller"
  /prioritizer/defer:
    post:
      tags:
      - "Recommender"
      summary: "Defer a requirement in the list."
      description: ""
      operationId: "defer_prioritized_issue"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Request object"
        required: true
        schema:
          $ref: "#/definitions/DeferRequirementRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DeferRequirementResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "application.controllers.recommendation_controller"
  /prioritizer/profile/delete:
    post:
      tags:
      - "Recommender"
      summary: "Delete a user's profile."
      description: ""
      operationId: "delete_profile"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Request object"
        required: true
        schema:
          $ref: "#/definitions/DeleteProfileRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DeleteProfileResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "application.controllers.recommendation_controller"
definitions:
  PrioritizedRecommendationsRequest:
    type: "object"
    properties:
      agentID:
        type: "string"
      assignee:
        type: "string"
      components:
        type: "array"
        items:
          type: "string"
      products:
        type: "array"
        items:
          type: "string"
      keywords:
        type: "array"
        items:
          type: "string"
    required:
      - agentID
      - assignee
      - components
      - products
      - keywords
  ChartRequest:
    type: "object"
    properties:
      agentID:
        type: "string"
      assignee:
        type: "string"
      components:
        type: "array"
        items:
          type: "string"
      products:
        type: "array"
        items:
          type: "string"
      keywords:
        type: "array"
        items:
          type: "string"
    required:
      - agentID
      - assignee
      - components
      - products
      - keywords
  LikeRequirementRequest:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      agentID:
        type: "string"
      assignee:
        type: "string"
      components:
        type: "array"
        items:
          type: "string"
      products:
        type: "array"
        items:
          type: "string"
      keywords:
        type: "array"
        items:
          type: "string"
    required:
      - id
      - agentID
      - assignee
      - components
      - products
      - keywords
  DeferRequirementRequest:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      agentID:
        type: "string"
      interval:
        type: "number"
        format: "float"
        minimum: 0
        maximum: 365
      assignee:
        type: "string"
      components:
        type: "array"
        items:
          type: "string"
      products:
        type: "array"
        items:
          type: "string"
      keywords:
        type: "array"
        items:
          type: "string"
    required:
      - id
      - agentID
      - interval
      - assignee
      - components
      - products
      - keywords
  DeleteProfileRequest:
    type: "object"
    properties:
      agentID:
        type: "string"
    required:
      - agentID
  PrioritizedRecommendationsResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      errorMessage:
        type: "string"
      rankedBugs:
        type: "array"
        items:
          type: "object"
          properties:
            id:
              type: "integer"
              format: "int64"
            summary:
              type: "string"
            product:
              type: "string"
            component:
              type: "string"
            priority:
              type: "number"
              format: "float"
              minimum: 0
              maximum: 1
            numberOfCC:
              type: "integer"
              format: "int32"
            milestone:
              type: "string"
            keywords:
              type: "array"
              items:
                type: "string"
    xml:
      name: "response"
  ChartResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      errorMessage:
        type: "string"
    xml:
      name: "response"
  LikeRequirementResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      errorMessage:
        type: "string"
    xml:
      name: "response"
  DeferRequirementResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      errorMessage:
        type: "string"
    xml:
      name: "response"
  DeleteProfileResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      errorMessage:
        type: "string"
    xml:
      name: "response"